requirements.txt
aiogram==3.3.0
pandas
python-docx
openpyxl
nest_asyncio